quantize and dequantize

+++ cags/quantize.py
@@ -13,20 +13,22 @@ def copy_not_exists(source, destination):
     shutil.copy(source, destination)
 
 
-def quantize(source, destination, iteration, sh_degree, device, **kwargs):
+def quantize(source, destination, iteration, sh_degree, device, quantize, **kwargs):
     copy_not_exists(os.path.join(source, "cfg_args"), os.path.join(destination, "cfg_args"))
     copy_not_exists(os.path.join(source, "cameras.json"), os.path.join(destination, "cameras.json"))
     input = os.path.join(source, "point_cloud", "iteration_" + str(iteration), "point_cloud.ply")
     output = os.path.join(destination, "point_cloud", "iteration_" + str(iteration), "point_cloud_quantized.ply")
-    shutil.rmtree(os.path.join(destination, "point_cloud", "iteration_" + str(iteration)))
-    os.makedirs(os.path.join(destination, "point_cloud", "iteration_" + str(iteration)), exist_ok=True)
     gaussians = GaussianModel(sh_degree).to(device)
     gaussians.load_ply(input)
     quantizer = ScalableQuantizer(gaussians, **kwargs)
-    gaussians = quantizer.save_quantized(output)
-    gaussians = quantizer.load_quantized(output)
-    output = os.path.join(destination, "point_cloud", "iteration_" + str(iteration), "point_cloud.ply")
-    gaussians.save_ply(output)
+    if quantize:
+        shutil.rmtree(os.path.join(destination, "point_cloud", "iteration_" + str(iteration)))
+        os.makedirs(os.path.join(destination, "point_cloud", "iteration_" + str(iteration)), exist_ok=True)
+        gaussians = quantizer.save_quantized(output)
+    else:
+        gaussians = quantizer.load_quantized(output)
+        output = os.path.join(destination, "point_cloud", "iteration_" + str(iteration), "point_cloud.ply")
+        gaussians.save_ply(output)
 
 
 if __name__ == "__main__":
@@ -38,8 +40,9 @@ if __name__ == "__main__":
     parser.add_argument("--sh_degree", type=int, default=3)
     parser.add_argument("--device", type=str, default="cuda")
     parser.add_argument("-o", "--option", default=[], action='append', type=str)
+    parser.add_argument("--dequantize", action='store_true')
     args = parser.parse_args()
     configs = {o.split("=", 1)[0]: eval(o.split("=", 1)[1]) for o in args.option}
     quantize(
         source=args.source, destination=args.destination, iteration=args.iteration, sh_degree=args.sh_degree,
-        device=args.device, **configs)
+        device=args.device, quantize=not args.dequantize, **configs)
